+++ src/features/generatingBingoCards.js
@@ -1,7 +1,7 @@
 // Generating Bingo Cards
 
 class BingoCard {
-  constructor (B, I, N, G, O) {
+  constructor(B, I, N, G, O) {
     this.B = B
     this.I = I
     this.N = N
@@ -11,7 +11,7 @@ class BingoCard {
 }
 
 class BingoCardGenerator {
-  getBingoCard () {
+  getBingoCard() {
     const B = this._fillArray(1, 15)
     const I = this._fillArray(16, 30)
     const N = this._fillArray(31, 45, 4)
@@ -21,17 +21,17 @@ class BingoCardGenerator {
     return new BingoCard(B, I, N, G, O)
   }
 
-  _randomNumberGeneratorBetweenBounds (lowerBound, upperBound) {
+  _randomNumberGeneratorBetweenBounds(lowerBound, upperBound) {
     return Math.floor(Math.random() * (upperBound - lowerBound) + lowerBound)
   }
 
-  _fillArray (lowerBound, upperBound, limit = 5) {
+  _fillArray(lowerBound, upperBound, limit = 5) {
     const aux = []
     for (let i = 0; i < limit; i++)
       aux.push(this._randomNumberGeneratorBetweenBounds(lowerBound, upperBound))
 
     return aux
-  } 
+  }
 }
 
 module.exports = {