+++ changelog.md
@@ -7,6 +7,13 @@
   - Calling Bingo Numbers section
 
 ## Version 0.2.0
-- Implemented.
+- Implemented:
   - Generating Bingo Cards
   - Updated documentation
+
+## Version 0.3.0
+- Implemented:
+  - Checking Bingo Cards
+  - Fixed changelog
+  - Updated documentation
+  - Prettifying files